+++ android/app/src/main/kotlin/deckers/thibault/aves/metadata/MediaMetadataRetrieverHelper.kt
@@ -15,7 +15,6 @@ object MediaMetadataRetrieverHelper {
         MediaMetadataRetriever.METADATA_KEY_ARTIST to "Artist",
         MediaMetadataRetriever.METADATA_KEY_AUTHOR to "Author",
         MediaMetadataRetriever.METADATA_KEY_BITRATE to "Bitrate",
-        MediaMetadataRetriever.METADATA_KEY_CAPTURE_FRAMERATE to "Capture Framerate",
         MediaMetadataRetriever.METADATA_KEY_CD_TRACK_NUMBER to "CD Track Number",
         MediaMetadataRetriever.METADATA_KEY_COMPILATION to "Compilation",
         MediaMetadataRetriever.METADATA_KEY_COMPOSER to "Composer",
@@ -35,6 +34,9 @@ object MediaMetadataRetrieverHelper {
         MediaMetadataRetriever.METADATA_KEY_WRITER to "Writer",
         MediaMetadataRetriever.METADATA_KEY_YEAR to "Year",
     ).apply {
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
+            put(MediaMetadataRetriever.METADATA_KEY_CAPTURE_FRAMERATE, "Capture Framerate")
+        }
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.P) {
             putAll(
                 hashMapOf(